Playground
This is a dev feature that put independent UI components together for view and testing.
Once you're done with a UI component, put it into the body of the playground page, and other people can see and reuse your component easily.
This is to be removed upon release, so you should not export anything from this page except for its own routing. On the other words, remove playground should not bring breaking change.